refactor(edda-ledger): seal sqlite_store module and rename DepRow (GH-378)
Move remaining public types (BundleRow, DeviceTokenRow, DecideSnapshotRow,
SuggestionRow, TaskBriefRow, ImportParams) from sqlite_store/types.rs to
domain.rs. Rename DepRow to DependencyEdge with backwards-compatible alias.

Change sqlite_store module visibility to pub(crate) — no downstream crate
can reach sqlite_store::* directly anymore. All types are re-exported
from lib.rs via the domain module.

Phase 3 of 3: completes the domain abstraction boundary.
