+++ crates/edda-ledger/src/domain.rs
@@ -124,3 +124,105 @@ pub struct ExecutionLinked {
     pub token_out: Option<u64>,
     pub latency_ms: Option<u64>,
 }
+
+// ── Types moved from sqlite_store/types.rs ──────────────────────────
+
+/// A review bundle row.
+#[derive(Debug, Clone)]
+pub struct BundleRow {
+    pub event_id: String,
+    pub bundle_id: String,
+    pub status: String,
+    pub risk_level: String,
+    pub total_added: i64,
+    pub total_deleted: i64,
+    pub files_changed: i64,
+    pub tests_passed: i64,
+    pub tests_failed: i64,
+    pub suggested_action: String,
+    pub branch: String,
+    pub created_at: String,
+}
+
+/// A dependency edge between two decision keys.
+#[derive(Debug, Clone)]
+pub struct DependencyEdge {
+    pub source_key: String,
+    pub target_key: String,
+    pub dep_type: String,
+    pub created_event: Option<String>,
+    pub created_at: String,
+}
+
+/// Parameters for inserting an imported decision from another project.
+pub struct ImportParams<'a> {
+    pub event: &'a edda_core::types::Event,
+    pub key: &'a str,
+    pub value: &'a str,
+    pub reason: &'a str,
+    pub domain: &'a str,
+    pub scope: &'a str,
+    pub source_project_id: &'a str,
+    pub source_event_id: &'a str,
+    pub is_active: bool,
+}
+
+/// A task brief row.
+#[derive(Debug, Clone)]
+pub struct TaskBriefRow {
+    pub task_id: String,
+    pub intake_event_id: String,
+    pub title: String,
+    pub intent: edda_core::types::TaskBriefIntent,
+    pub source_url: String,
+    pub status: edda_core::types::TaskBriefStatus,
+    pub branch: String,
+    pub iterations: i64,
+    pub artifacts: String,
+    pub decisions: String,
+    pub last_feedback: Option<String>,
+    pub created_at: String,
+    pub updated_at: String,
+}
+
+/// A device token row.
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+pub struct DeviceTokenRow {
+    pub token_hash: String,
+    pub device_name: String,
+    pub paired_at: String,
+    pub paired_from_ip: String,
+    pub revoked_at: Option<String>,
+    pub pair_event_id: String,
+    pub revoke_event_id: Option<String>,
+}
+
+/// A decide snapshot row.
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+pub struct DecideSnapshotRow {
+    pub event_id: String,
+    pub context_hash: String,
+    pub engine_version: String,
+    pub schema_version: String,
+    pub redaction_level: String,
+    pub village_id: Option<String>,
+    pub cycle_id: Option<String>,
+    pub has_blobs: bool,
+    pub created_at: String,
+}
+
+/// A suggestion row.
+#[derive(Debug, Clone)]
+pub struct SuggestionRow {
+    pub id: String,
+    pub event_type: String,
+    pub source_layer: String,
+    pub source_refs: String,
+    pub summary: String,
+    pub suggested_because: String,
+    pub detail: String,
+    pub tags: String,
+    pub status: String,
+    pub created_at: String,
+    pub reviewed_at: Option<String>,
+}
