+++ crates/edda-ledger/src/lib.rs
@@ -5,7 +5,7 @@ pub mod domain;
 pub mod ledger;
 pub mod lock;
 pub mod paths;
-pub mod sqlite_store;
+pub(crate) mod sqlite_store;
 pub mod sync;
 pub mod tombstone;
 pub mod view;
@@ -17,15 +17,13 @@ pub use blob_store::{
     SNAPSHOT_BLOB_THRESHOLD,
 };
 pub use domain::{
-    ChainEntryView, DayCount, DetectedPattern, DomainCount, ExecutionLinked, OutcomeMetrics,
-    PatternDetectionResult, PatternType, VillageStats, VillageStatsPeriod,
+    BundleRow, ChainEntryView, DayCount, DecideSnapshotRow, DependencyEdge, DetectedPattern,
+    DeviceTokenRow, DomainCount, ExecutionLinked, ImportParams, OutcomeMetrics,
+    PatternDetectionResult, PatternType, SuggestionRow, TaskBriefRow, VillageStats,
+    VillageStatsPeriod,
 };
 pub use ledger::Ledger;
 pub use lock::WorkspaceLock;
 pub use paths::EddaPaths;
-pub use sqlite_store::{
-    BundleRow, ChainEntry, DecideSnapshotRow, DecisionRow, DepRow, DeviceTokenRow, ImportParams,
-    SuggestionRow, TaskBriefRow,
-};
 pub use tombstone::{append_tombstone, list_tombstones, make_tombstone, DeleteReason, Tombstone};
 pub use view::DecisionView;
